Move vault auth logic from SBT to app code.
