+++ project/src/main/scala/org/broadinstitute/clio/integrationtest/ClioIntegrationTestSettings.scala
@@ -2,39 +2,15 @@ package org.broadinstitute.clio.integrationtest
 
 import org.broadinstitute.clio.sbt.{Compilation, Dependencies, Versioning}
 
-import com.bettercloud.vault.{Vault, VaultConfig}
-import com.google.auth.oauth2.ServiceAccountCredentials
 import com.lucidchart.sbt.scalafmt.ScalafmtCorePlugin
 import sbt._
 import sbt.Def.{Initialize, Setting}
 import sbt.Keys._
 
-import scala.collection.JavaConverters._
-
 /**
   * Settings for running our integration tests.
   */
 object ClioIntegrationTestSettings {
-
-  /** URL of vault server to use when getting bearer tokens for service accounts. */
-  private val vaultUrl = "https://clotho.broadinstitute.org:8200/"
-
-  /** List of possible token-file locations, in order of preference. */
-  private val vaultTokenFiles = Seq(
-    new File("/etc/vault-token-dsde"),
-    new File(s"${System.getProperty("user.home")}/.vault-token")
-  )
-
-  /**
-    * Path within Vault to the service account info Jenkins should use when talking to Clio.
-    */
-  private val vaultPath = "secret/dsde/gotc/dev/clio/clio-test.json"
-
-  /** Scopes needed from Google to get past Clio's auth proxy. */
-  private val authScopes = Seq(
-    "https://www.googleapis.com/auth/userinfo.profile",
-    "https://www.googleapis.com/auth/userinfo.email"
-  )
 
   /** Directory to which all container logs will be sent during Dockerized integration tests. */
   lazy val logTarget: Initialize[File] = Def.setting {
@@ -61,46 +37,6 @@ object ClioIntegrationTestSettings {
     clioLog
   }
 
-  /** Task to get a bearer token for use in integration tests. */
-  lazy val getBearerToken: Initialize[Task[String]] = Def.task {
-    if (sys.env.get("JENKINS_URL").isEmpty) {
-      // If we're not running in Jenkins, we can just shell out to gcloud
-      "gcloud auth print-access-token".!!.stripLineEnd
-    } else {
-      // Otherwise, we have to go through Vault to get a service account's
-      // info, through which we can get an access token
-      val vaultToken = sys.env
-        .get("VAULT_TOKEN")
-        .orElse(vaultTokenFiles.find(_.exists).map(IO.read(_).stripLineEnd))
-        .getOrElse {
-          sys.error(
-            "Vault token not given or found on filesystem, can't get bearer token!"
-          )
-        }
-
-      val vaultConfig = new VaultConfig()
-        .address(vaultUrl)
-        .token(vaultToken)
-        .build()
-
-      val vaultDriver = new Vault(vaultConfig)
-      val accountInfo = vaultDriver.logical().read(vaultPath).getData
-
-      val credential =
-        ServiceAccountCredentials.fromPkcs8(
-          accountInfo.get("client_id"),
-          accountInfo.get("client_email"),
-          accountInfo.get("private_key"),
-          accountInfo.get("private_key_id"),
-          authScopes.asJava,
-          null,
-          new URI(accountInfo.get("token_uri"))
-        )
-
-      credential.refreshAccessToken().getTokenValue
-    }
-  }
-
   /** Environment variables to inject when running integration tests. */
   lazy val itEnvVars: Initialize[Task[Map[String, String]]] = Def.task {
     val logDir = logTarget.value
@@ -112,8 +48,7 @@ object ClioIntegrationTestSettings {
       "ELASTICSEARCH_DOCKER_TAG" -> Dependencies.ElasticsearchVersion,
       "LOG_DIR" -> logDir.getAbsolutePath,
       "CLIO_LOG_FILE" -> clioLog.getAbsolutePath,
-      "CONF_DIR" -> confDir.getAbsolutePath,
-      "BEARER_TOKEN" -> getBearerToken.value
+      "CONF_DIR" -> confDir.getAbsolutePath
     )
   }
